+++ src/main/java/chess/dao/dto/LatestChessBoardDto.java
@@ -1,4 +1,4 @@
-package chess.repository.dto;
+package chess.dao.dto;
 
 import chess.model.board.ChessBoard;
 import chess.model.board.Turn;
