+++ src/main/java/chess/dao/exception/DataAccessException.java
@@ -1,4 +1,4 @@
-package chess.repository.exception;
+package chess.dao.exception;
 
 public class DataAccessException extends RuntimeException {
     public DataAccessException(String message) {
