+++ src/main/java/chess/service/ChessGameService.java
@@ -8,13 +8,13 @@ import chess.model.piece.Blank;
 import chess.model.piece.Piece;
 import chess.model.piece.Side;
 import chess.model.position.Position;
-import chess.repository.dao.ChessBoardDao;
-import chess.repository.dao.PieceDao;
-import chess.repository.dto.GameResultDto;
-import chess.repository.dto.LatestChessBoardDto;
-import chess.repository.dto.NewChessBoardDto;
-import chess.repository.exception.DataAccessException;
-import chess.repository.util.MySqlConnector;
+import chess.dao.ChessBoardDao;
+import chess.dao.PieceDao;
+import chess.dao.dto.GameResultDto;
+import chess.dao.dto.LatestChessBoardDto;
+import chess.dao.dto.NewChessBoardDto;
+import chess.dao.exception.DataAccessException;
+import chess.dao.util.MySqlConnector;
 import chess.view.input.MoveArguments;
 
 import java.util.List;
