+++ src/test/java/chess/dao/util/MySqlConnectorTest.java
@@ -1,4 +1,4 @@
-package chess.repository.util;
+package chess.dao.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
