+++ src/test/java/chess/dao/util/TestMySqlConnector.java
@@ -1,4 +1,4 @@
-package chess.repository.util;
+package chess.dao.util;
 
 public class TestMySqlConnector extends MySqlConnector {
     public static final TestMySqlConnector INSTANCE = new TestMySqlConnector();
