+++ src/test/java/chess/service/ChessGameServiceTest.java
@@ -9,11 +9,11 @@ import chess.model.piece.Side;
 import chess.model.position.File;
 import chess.model.position.Position;
 import chess.model.position.Rank;
-import chess.repository.DataBaseCleaner;
-import chess.repository.dao.ChessBoardDao;
-import chess.repository.dao.PieceDao;
-import chess.repository.dto.GameResultDto;
-import chess.repository.util.TestMySqlConnector;
+import chess.dao.DataBaseCleaner;
+import chess.dao.ChessBoardDao;
+import chess.dao.PieceDao;
+import chess.dao.dto.GameResultDto;
+import chess.dao.util.TestMySqlConnector;
 import chess.view.input.MoveArguments;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
